Refactor Interface class
Currently this class is only used for modular compilation, but it is
also useful to implement interface lookup (coming soon), i.e.
indexing a list of interfaces using a bit-vector.  The new (more
general) class structure will allow this.  It also generates better
interface names that the previous approach.  The disadvantage is that
defining instances of this class is a bit more involved, but this may
be fixable.  Anyway, the common case is to use generic (automatic)
deriving for this class.