Tweak to simulate function
The 'simulate' function doesn't allow external inputs to be provided,
yet allows simulation of a circuit the containing external inputs.
This leads to nasty errors when the inputs cannout be found. Fix this
by requiring the circuit to have no inputs or outputs at the type
level.  Later we can provide a lower-level version of this function
where external inputs can be provided.