Apply fix from @Swire42 & avoid resulting z3 issue
@Swire42 found a bug in Blarney's state transition chaining function.
Fixing this bug led to a surprising performance regression in z3. This
commit fixes the bug and avoids the performance regression by avoiding the
use of a recursive chaining function in the generated SMT. In addition,
z3 performance now appears to be much higher than it was previously.

Note: I only avoid the performance regression for depth-bounded
checking. We should also avoid it when using induction. I intend to look
at this soon, as well as the alternative SMT backend from @Swire42.